ScanToTextPDF/requirements.txt
PyPDF2
Pillow
pytesseract
pdf2image
img2pdf
reportlab